# src-app/ui/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "",
    "version": ""
  },
  "paths": {
...
    "/api/mcp/system-servers/{id}/groups": {
      "get": {
        "tags": [
          "MCP Servers - System"
        ],
        "summary": "Get server's assigned groups",
        "description": "Get groups assigned to an MCP server",
        "operationId": "McpServerSystem.getServerGroups",
        "responses": {
          "403": {
            "description": "Forbidden - Missing required permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionError"
                },
                "example": {
                  "details": {
                    "required_permissions": [
                      {
                        "description": "View system MCP servers",
                        "name": "McpServersAdminRead",
                        "value": "mcp_servers_admin::read"
                      }
                    ]
                  },
                  "error": "Missing required permission: mcp_servers_admin::read",
                  "error_code": "INSUFFICIENT_PERMISSIONS"
                }
              }
            }
          },
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string",
                    "format": "uuid"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "MCP Servers - System"
        ],
        "summary": "Assign server to groups",
        "description": "Assign MCP server to groups (replaces all assignments)",
        "operationId": "McpServerSystem.assignServerToGroups",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ServerGroupsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "403": {
            "description": "Forbidden - Missing required permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionError"
                },
                "example": {
                  "details": {
                    "required_permissions": [
                      {
                        "description": "Edit system MCP servers and manage group assignments",
                        "name": "McpServersAdminEdit",
                        "value": "mcp_servers_admin::edit"
                      }
                    ]
                  },
                  "error": "Missing required permission: mcp_servers_admin::edit",
                  "error_code": "INSUFFICIENT_PERMISSIONS"
                }
              }
            }
          },
          "204": {
            "description": "Server assigned successfully"
          },
          "400": {
            "description": "Bad request - only system servers can be assigned to groups"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Server not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/mcp/system-servers/{id}/groups/{group_id}": {
      "delete": {
        "tags": [
          "MCP Servers - System"
        ],
        "summary": "Remove server from group",
        "description": "Remove an MCP server from a specific group",
        "operationId": "McpServerSystem.removeServerFromGroup",
        "responses": {
          "403": {
            "description": "Forbidden - Missing required permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PermissionError"
                },
                "example": {
                  "details": {
                    "required_permissions": [
                      {
                        "description": "Edit system MCP servers and manage group assignments",
                        "name": "McpServersAdminEdit",
                        "value": "mcp_servers_admin::edit"
                      }
                    ]
                  },
                  "error": "Missing required permission: mcp_servers_admin::edit",
                  "error_code": "INSUFFICIENT_PERMISSIONS"
                }
              }
            }
          },
          "204": {
            "description": "Server removed successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Server assignment not found"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PermissionDetail": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "value",
          "description"
        ]
      },
      "PermissionError": {
        "description": "403 Forbidden response for missing permissions",
        "type": "object",
        "properties": {
          "details": {
            "$ref": "#/components/schemas/PermissionErrorDetails"
          },
          "error": {
            "type": "string"
          },
          "error_code": {
            "type": "string"
          }
        },
        "required": [
          "error",
          "error_code",
          "details"
        ]
      },
      "PermissionErrorDetails": {
        "type": "object",
        "properties": {
          "required_permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PermissionDetail"
            }
          }
        },
        "required": [
          "required_permissions"
        ]
      },
...
      "ServerGroupsRequest": {
        "type": "object",
        "properties": {
          "group_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          }
        },
        "required": [
          "group_ids"
        ]
      },
...
    }
  }
}